requirements.txt
keras==2.14.0
matplotlib==3.6.2
numpy==1.23.5
Pillow==9.3.0
Pillow==10.1.0
scikit_learn==1.3.1
seaborn==0.13.0
tensorflow==2.14.0
